Added Ability for Fighter
Added the ability for fighter to expend its rage to strike the enemy for increase damage. Also fixed the mana bar to be dynamic at runtime instead of always showing a max of 100. Fixed the same issue with the healthbar. For some reason we don't have to call setHealth() so we should check that out.